refactor: clean up more
