+++ test/fast/commands/branch/branch_delete.test.ts
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { allBranchesWithMeta } from '../../../../src/wrapper-classes/metadata_ref';
+import { initContext } from '../../../../src/lib/context';
 import { allScenes } from '../../../lib/scenes/all_scenes';
 import { configureTest } from '../../../lib/utils/configure_test';
 import { expectBranches } from '../../../lib/utils/expect_branches';
@@ -22,23 +22,9 @@ for (const scene of allScenes) {
 
       expectBranches(scene.repo, 'main');
 
-      expect(allBranchesWithMeta().find((b) => b === branchName)).to.be
-        .undefined;
-    });
-
-    it('Can run branch delete on a branch not created/tracked by Graphite', () => {
-      const branchName = 'a';
-
-      scene.repo.createAndCheckoutBranch(branchName);
-      scene.repo.createChangeAndCommit('2', '2');
-
-      scene.repo.checkoutBranch('main');
-      scene.repo.execCliCommandAndGetOutput(`bdl "${branchName}" -f -q`);
-
-      expectBranches(scene.repo, 'main');
-
-      expect(allBranchesWithMeta().find((b) => b === branchName)).to.be
-        .undefined;
+      expect(
+        initContext().metaCache.allBranchNames.find((b) => b === branchName)
+      ).to.be.undefined;
     });
   });
 }
